Calculate the necessary for one simulation.

Calculate the necessary for one simulation.

Usage

basic_simulation(
  F_mat_W,
  L_mat_W,
  F_mat_Y,
  L_mat_Y,
  cov_mat_fs,
  cov_mat_rf,
  pi_unit,
  theta_w,
  theta_y,
  tau,
  no_noise,
  rho_agg,
  rho_cross,
  T_0,
  Z_fit,
  test = FALSE,
  S = 300
)

Arguments

F_mat_W

a nxT two-way fixed effects matrix for W (endogenous treatment).

L_mat_W

a nxT generalized fixed effects (low-rank) matrix for W (endogenous treatment).

F_mat_Y

a nxT two-way fixed effects matrix for Y (outcome).

L_mat_Y

a nxT generalized fixed effects (low-rank) matrix for Y (outcome).

cov_mat_fs

a covarience matrix for W (endogenous treatment) noise.

cov_mat_rf

a covarience matrix for a part of Y (outcome) noise.

pi_unit

a nx1 vector of exposure.

theta_w

an exposure to unobserved confounder for W (endogenous treatment).

theta_y

an exposure to unobserved confounder for Y (outcome).

tau

the original point estimate of tau. It is used for data generation.

no_noise

If true, the noise is not added. We do not add noise for table 3.

rho_agg

the share of Z_t during H_t generation.

rho_cross

the share of W (endogenous treatment) noise in Y noise.

T_0

the size of the learning period.

Z_fit

the fitted ARIMA model to the data Z (instrument).

test

If true, compute the tests for the coverage rates.

S

if 'test' == TRUE, the number of simulations for standard error estimation.

Value

A list with simulation results.