Compute the robust estimator for tau.

Compute the robust estimator for tau.

robust_estimate(
  Y_mat_or,
  W_mat_or,
  Z,
  unit_covariates = NULL,
  time_covariates = NULL,
  index_sub = NULL,
  T_0 = NULL,
  unit_names = NULL,
  time_column = TRUE,
  seed = 1234
)

Arguments

Y_mat_or

an nxT matrix of outcome.

W_mat_or

an nxT matrix of endogenous treatment.

Z

an nx1 vector of an aggregate instrument.

index_sub

an nx1 vector with binary coordinates. If one of the coordinates is TRUE, we include this state in the estimation.

T_0

the size of the learning period.

unit_names

an nx1 vector with unit names. It is used only in plot_1 function.

time_column

If true, the units are in the rows and the time periods are in the columns in Y_mat_or and W_mat_or.

seed,

seed to set.

D

an nx1 vector of exposure.

Value

A robust estimator with result, attached as an attribute. It is an output of basic_analysis function.