Perform the simulations and present the results.

Perform the simulations and present the results.

Usage

simulations(
  Y_mat_or,
  W_mat_or,
  Z,
  share_t = 1/3,
  share_rank = 1/3,
  no_noise = FALSE,
  rho_agg = 0.5,
  rho_theta_w = 0.2,
  rho_theta_y = 0.3,
  B = 1000,
  S = 300,
  test = FALSE,
  return_table = FALSE,
  K = 300,
  deg = 4,
  height = 9 * 0.75,
  width = 16 * 0.75,
  save_pdf = FALSE,
  file_plot = "density_plot",
  draw_plot = TRUE,
  save_sim = FALSE,
  file_sim = "simulation_result",
  seed = 1234
)

Arguments

Y_mat_or

an nxT matrix of outcome.

W_mat_or

an nxT matrix of endogenous treatment.

Z

n nx1 vector of an aggregate instrument.

share_t

the share of the learning period.

share_rank

the share of the rank of generalized fixed effects matrix in total rank.

no_noise

If true, the noise is not added. We do not add noise for table 3.

rho_agg

the share of Z_t during H_t generation.

rho_theta_w

the exposure of W (endogenous treatment) to unobserved confounder.

rho_theta_y

the exposure of Y (outcome) to unobserved confounder.

B

number of replications for each design of simulation.

S

if 'test' == TRUE, the number of simulations for standard error estimation.

test

If true, compute the tests for the coverage rates.

return_table

If true, instead of printing table(s) the function returns it(them).

K

number of domain splits. It is used in density plot.

deg

degrees of freedom for natural cubic splines. It is used in density plot.

height

the height of the plot.

width

the width of the plot.

save_pdf

If true, the 2 density plots for Original and Robust estimators are saved.

file_plot

If save_pdf is true, string giving the file path (including the file name).

draw_plot

If draw_plot is true and save_pdf is false, the 2 density plots are drawn in one chart.

save_sim

If true, the function saves the simulation.

file_sim

If save_sim is true, string giving the file path (including the file name).

seed

seed to set.